refactor: extract prompt type to module
