+++ src/types/icrc-standards.ts
@@ -24,14 +24,6 @@ export const IcrcMethodSchema = z.enum([
   ICRC49_CALL_CANISTER
 ]);
 
-export const IcrcApproveMethodSchema = z.enum([
-  ICRC25_REQUEST_PERMISSIONS,
-  ICRC27_ACCOUNTS,
-  ICRC49_CALL_CANISTER
-]);
-
-export type IcrcApproveMethod = z.infer<typeof IcrcApproveMethodSchema>;
-
 export const IcrcScopedMethodSchema = IcrcMethodSchema.extract([
   ICRC27_ACCOUNTS,
   ICRC49_CALL_CANISTER
